# frontend/openapi.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "Village API",
    "version": "0.1.0"
  },
  "paths": {
...
    "\/user\/create": {
      "post": {
        "tags": [
          "users"
        ],
        "summary": "Get Or Create User",
        "operationId": "get_or_create_user",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application\/json": {
                "schema": {
                  "$ref": "#\/components\/schemas\/UserWithWorkspaces"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "\/user\/get": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Get Current User",
        "description": "Get a user.",
        "operationId": "get_current_user",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application\/json": {
                "schema": {
                  "$ref": "#\/components\/schemas\/UserWithWorkspaces"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "\/workspace\/propose_id": {
      "get": {
        "tags": [
          "workspaces"
        ],
        "summary": "Propose Workspace Id",
        "description": "Propose a workspace ID based on the name.",
        "operationId": "propose_workspace_id",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            },
            "name": "name",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application\/json": {
                "schema": {
                  "title": "Response Propose Workspace Id",
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application\/json": {
                "schema": {
                  "$ref": "#\/components\/schemas\/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "\/workspace\/check_id": {
      "get": {
        "tags": [
          "workspaces"
        ],
        "summary": "Check Workspace Id",
        "description": "Check if a script ID is available.",
        "operationId": "check_workspace_id",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Id",
              "type": "string"
            },
            "name": "id",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application\/json": {
                "schema": {
                  "title": "Response Check Workspace Id",
                  "type": "boolean"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application\/json": {
                "schema": {
                  "$ref": "#\/components\/schemas\/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "\/workspace\/create": {
      "post": {
        "tags": [
          "workspaces"
        ],
        "summary": "Create Workspace",
        "description": "Create a workspace.",
        "operationId": "create_workspace",
        "requestBody": {
          "content": {
            "application\/json": {
              "schema": {
                "$ref": "#\/components\/schemas\/CreateWorkspaceInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application\/json": {
                "schema": {
                  "$ref": "#\/components\/schemas\/Workspace"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application\/json": {
                "schema": {
                  "$ref": "#\/components\/schemas\/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "\/workspace\/get": {
      "get": {
        "tags": [
          "workspaces"
        ],
        "summary": "Get Workspace",
        "description": "Get a workspace.",
        "operationId": "get_workspace",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Workspace Id",
              "type": "string"
            },
            "name": "workspace_id",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application\/json": {
                "schema": {
                  "$ref": "#\/components\/schemas\/Workspace"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application\/json": {
                "schema": {
                  "$ref": "#\/components\/schemas\/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "\/workspace\/list": {
      "get": {
        "tags": [
          "workspaces"
        ],
        "summary": "List User Workspaces",
        "description": "List all workspaces.",
        "operationId": "list_user_workspaces",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application\/json": {
                "schema": {
                  "title": "Response List User Workspaces",
                  "type": "array",
                  "items": {
                    "$ref": "#\/components\/schemas\/Workspace"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "\/workspace\/workspace\/list_users": {
      "get": {
        "tags": [
          "workspaces"
        ],
        "summary": "List Workspace Users",
        "description": "List all users in a workspace.",
        "operationId": "list_workspace_users",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Workspace Id",
              "type": "string"
            },
            "name": "workspace_id",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application\/json": {
                "schema": {
                  "title": "Response List Workspace Users",
                  "type": "array",
                  "items": {
                    "$ref": "#\/components\/schemas\/WorkspaceUsers"
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application\/json": {
                "schema": {
                  "$ref": "#\/components\/schemas\/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "\/workspace\/set_default": {
      "post": {
        "tags": [
          "workspaces"
        ],
        "summary": "Set Default Workspace",
        "description": "Set default workspace.",
        "operationId": "set_default_workspace",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Workspace Id",
              "type": "string"
            },
            "name": "workspace_id",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application\/json": {
                "schema": {
                  "title": "Response Set Default Workspace",
                  "type": "boolean"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application\/json": {
                "schema": {
                  "$ref": "#\/components\/schemas\/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}